feat: patch tag previews
Applied patch tag previews.

Some of the changes of the 'focusonclick' patch where undone, because of
the need of the function 'motionnotify' that was removed by the patch.

Some changes where made in the function 'motionnotify' to keep the
'focusonclick' patch full functionality.

The struct Monitor where manually recreate to keep the members gappih,
gappiv, gappoh and gappov from the patch 'cfacts vanitygaps combo'

The member 'tagmap' of the struct Monitor was changed to use a constant
length of 9, there as a comment with "FIXME" left in there because the
constant value may cause problems in the future. (but this is for future
me to worry about, so good luck future me... ¯\_(ツ)_/¯).

The 'updatepreview' function was changed to take into account the values
from the 'barpadding' patch.

source: https://dwm.suckless.org/patches/tag-previews/dwm-tag-preview-6.2.diff